117 of 167 files compile
svn path=/trunk/KDE/kdepim/; revision=461055

+++ kmsender.cpp
@@ -1128,7 +1128,7 @@ void KMSendSMTP::doFinish() {
 void KMSendSMTP::dataReq(KIO::Job *, QByteArray &array)
 {
   // Send it by 32K chuncks
-  const int chunkSize = QMIN( mMessageLength - mMessageOffset, 32*1024 );
+  const int chunkSize = QMIN( mMessageLength - mMessageOffset, (unsigned int)32*1024 );
   if ( chunkSize > 0 ) {
     array.duplicate(mMessage.data() + mMessageOffset, chunkSize);
     mMessageOffset += chunkSize;
